requirements.txt
flask==3.0.3
flask-cors==4.0.1
Pillow==11.0.0
rembg==2.0.68
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
gunicorn==21.2.0
